+++ biz/webui/htdocs/src/js/req-data.js
@@ -15,13 +15,13 @@ var events = require('./events');
 var iframes = require('./iframes');
 var dataCenter = require('./data-center');
 var storage = require('./storage');
-var MockDialog = require('./mock-dialog');
 
 var TREE_ROW_HEIGHT = 24;
 var ROW_STYLE = { outline: 'none' };
 var columnState = {};
 var columnKeys = {};
 var CMD_RE = /^:dump\s+(\d{1,15})\s*$/;
+var BODY_FILTER = /(^\s*|\s+)(content|c|b|body):/i;
 var NOT_BOLD_RULES = {
   plugin: 1,
   pac: 1,
@@ -361,6 +361,7 @@ var ReqData = React.createClass({
     };
   },
   componentDidUpdate: function () {
+    this.isShownBtn && events.trigger('checkAtBottom');
     if (storage.get('disabledHNR') === '1') {
       return;
     }
@@ -479,11 +480,6 @@ var ReqData = React.createClass({
   componentDidMount: function () {
     var self = this;
     var timer;
-    events.on('showMockDialog', function(_, data) {
-      if (data) {
-        self.refs.mockDialog.show(data.item, data.type);
-      }
-    });
     events.on('hashFilterChange', function () {
       self.setState({});
     });
@@ -520,14 +516,21 @@ var ReqData = React.createClass({
     };
     self.container = $(ReactDOM.findDOMNode(self.refs.container));
     self.content = ReactDOM.findDOMNode(self.refs.content);
+    var clickedCount = 0;
     self.$content = $(self.content)
       .on('dblclick', 'tr', function (e) {
-        events.trigger('toggleDetailTab');
+        if (clickedCount > 2) {
+          events.trigger('toggleDetailTab');
+        }
       })
       .on('click', 'tr', function (e) {
         var id = this.getAttribute('data-id');
         if (id) {
-          dataCenter.lastSelectedDataId = id;
+          if (dataCenter.lastSelectedDataId != id) {
+            clickedCount = 0;
+            dataCenter.lastSelectedDataId = id;
+          }
+          ++clickedCount;
           var item = self.props.modal.getItem(id);
           self.onClick(e, item);
         }
@@ -606,6 +609,25 @@ var ReqData = React.createClass({
     };
     importRemoteUrl();
     $(window).on('hashchange', importRemoteUrl);
+    var backBtn = $(ReactDOM.findDOMNode(self.refs.backBtn));
+    var hideBackBtn = function() {
+      if (self.isShownBtn) {
+        self.isShownBtn = false;
+        backBtn.hide();
+      }
+    };
+    self.hideBackBtn = hideBackBtn;
+    self.isShownBtn = false;
+    events.on('toggleBackToBottomBtn', function(_, show) {
+      if (show) {
+        if (!self.isShownBtn) {
+          self.isShownBtn = true;
+          backBtn.show();
+        }
+      } else {
+        hideBackBtn();
+      }
+    });
   },
   onDragStart: function (e) {
     var target = $(e.target).closest('.w-req-data-item');
@@ -882,7 +904,7 @@ var ReqData = React.createClass({
       self.collapseAll(treeId);
       break;
     case 'Mock':
-      self.refs.mockDialog.show(item);
+      events.trigger('showMockDialog', {item: item});
       break;
     }
   },
@@ -1087,7 +1109,8 @@ var ReqData = React.createClass({
       dataCenter.getNetworkMenus(),
       treeItem.hide ? 8 : 9,
       disabled,
-      treeId
+      treeId,
+      item && item.url
     );
     var height = (treeItem.hide ? 310 : 340) - (pluginItem.hide ? 30 : 0) - (mockItem.hide ? 30 : 0);
     pluginItem.maxHeight = height;
@@ -1101,9 +1124,11 @@ var ReqData = React.createClass({
   },
   onFilterChange: function (keyword) {
     var self = this;
-    self.props.modal.search(keyword);
+    var filterBody = BODY_FILTER.test(keyword);
     clearTimeout(self.networkStateChangeTimer);
+    !filterBody && self.props.modal.search(keyword);
     self.networkStateChangeTimer = setTimeout(function () {
+      filterBody && self.props.modal.search(keyword);
       self.setState({ filterText: keyword }, self.updateList);
       events.trigger('networkStateChange');
     }, 600);
@@ -1117,11 +1142,8 @@ var ReqData = React.createClass({
     this.refs.filterInput.clearFilterText();
   },
   autoRefresh: function () {
-    if (this.container) {
-      this.container.find(
-        '.ReactVirtualized__Grid:first'
-      ).scrollTop = 100000000;
-    }
+    this.container.find('.ReactVirtualized__Grid:first')[0].scrollTop = 100000000;
+    this.hideBackBtn();
   },
   orderBy: function (e) {
     var target = this.willResort && $(e.target).closest('th')[0];
@@ -1324,7 +1346,6 @@ var ReqData = React.createClass({
     return (
       <div className="fill w-req-data-con orient-vertical-box">
         <div
-          ref="wrapper"
           className="w-req-data-content fill orient-vertical-box"
           style={colStyle}
         >
@@ -1398,6 +1419,10 @@ var ReqData = React.createClass({
               }}
             </RV.AutoSizer>
           </div>
+          <div className="w-back-to-the-bottom" ref="backBtn" onClick={this.autoRefresh}>
+            <span className="glyphicon glyphicon-arrow-down" />
+            Back to the bottom
+          </div>
         </div>
         <FilterInput
           ref="filterInput"
@@ -1409,7 +1434,6 @@ var ReqData = React.createClass({
         <ContextMenu onClick={this.onClickContextMenu} ref="contextMenu" />
         <ContextMenu onClick={this.onClickHeadMenu} ref="headContextMenu" />
         <QRCodeDialog ref="qrcodeDialog" />
-        <MockDialog ref="mockDialog" />
       </div>
     );
   }